refactor: app 패키지를 mvc 패키지로 이동
