+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/HandlerRegistry.java
@@ -1,8 +1,6 @@
-package com.techcourse;
+package com.interface21.webmvc.servlet.mvc;
 
 import com.interface21.web.bind.annotation.RequestMethod;
-import com.interface21.webmvc.servlet.mvc.tobe.HandlerAdapter;
-import com.interface21.webmvc.servlet.mvc.tobe.HandlerMapping;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
